refactor: 프로젝트 레이어 및 패키지 구조에 맞춰 테스트 클래스 위치 재구성
