+++ app/src/test/java/woowacourse/shopping/data/remote/mock/ProductWebServerTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.data.remote.mock
 
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
@@ -7,9 +7,7 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.data.remote.mock.ProductWebServer
 import woowacourse.shopping.domain.repository.ProductRepository
-import woowacourse.shopping.data.remote.mock.MockProductRepositoryImpl
 
 class ProductWebServerTest {
     @Test
