+++ app/src/test/java/woowacourse/shopping/domain/model/ProductTest.kt
@@ -1,8 +1,7 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import woowacourse.shopping.domain.model.Product
 
 class ProductTest {
     @Test
