+++ app/src/test/java/woowacourse/shopping/testing/fakes/FakeRecentlyViewedProductRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.viewmodel.fakes
+package woowacourse.shopping.testing.fakes
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
